feat: refractor useage
