+++ components/Condition.tsx
