+++ components/Pattern.tsx
