+++ components/style/SortableCondition.css
