+++ components/style/SortableCondition.styl
