+++ components/utils/extractConditionConfig.ts
