+++ components/utils/extractPatternConfig.ts
