+++ components/utils/getPatternsChangeTreeData.ts
