+++ components/utils/isAllConditionItems.ts
