+++ components/utils/isAllNormalItems.ts
